Added OAM DMA support. Fixed bug with 32-bit int

+++ types.h
@@ -34,8 +34,8 @@ typedef unsigned char uint8;
 typedef signed char sint8;
 typedef unsigned short int uint16;
 typedef signed short int sint16;
-typedef unsigned long int uint32;
-typedef signed long int sint32;
+typedef unsigned int uint32;
+typedef signed int sint32;
 typedef unsigned long long uint64;
 typedef signed long long sint64;
